moved conversion_rates to features folder
