+++ packages/espressocash_app/lib/app/screens/authenticated/authenticated_flow_screen.dart
@@ -8,12 +8,12 @@ import 'package:provider/provider.dart';
 import '../../../core/accounts/bl/account.dart';
 import '../../../core/accounts/bl/accounts_bloc.dart';
 import '../../../core/balances/bl/balances_bloc.dart';
-import '../../../core/conversion_rates/bl/conversion_rates_bloc.dart';
-import '../../../core/conversion_rates/module.dart';
 import '../../../core/user_preferences.dart';
 import '../../../di.dart';
 import '../../../features/activities/module.dart';
 import '../../../features/backup_phrase/module.dart';
+import '../../../features/conversion_rates/conversion_rates.dart';
+import '../../../features/conversion_rates/module.dart';
 import '../../../features/favorite_tokens/module.dart';
 import '../../../features/incoming_single_link_payments/module.dart';
 import '../../../features/incoming_split_key_payments/module.dart';
