+++ packages/espressocash_app/lib/app/screens/authenticated/investments/components/context_ext.dart
@@ -3,9 +3,9 @@ import 'package:dfunc/dfunc.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 
-import '../../../../../core/balances/presentation/watch_balance.dart';
 import '../../../../../core/tokens/token.dart';
 import '../../../../../core/user_preferences.dart';
+import '../../../../../features/balances/presentation/watch_balance.dart';
 
 extension InvestmentExt on BuildContext {
   bool userHasInvestments() => watchUserTotalFiatBalance(
