+++ packages/espressocash_app/lib/app/screens/authenticated/wallet_flow/wallet_flow_screen.dart
@@ -4,9 +4,9 @@ import 'package:flutter/material.dart';
 
 import '../../../../core/amount.dart';
 import '../../../../core/currency.dart';
-import '../../../../core/presentation/format_amount.dart';
 import '../../../../core/tokens/token_list.dart';
 import '../../../../di.dart';
+import '../../../../features/balances/presentation/format_amount.dart';
 import '../../../../features/incoming_single_link_payments/widgets/extensions.dart';
 import '../../../../features/outgoing_direct_payments/widgets/extensions.dart';
 import '../../../../features/outgoing_split_key_payments/widgets/extensions.dart';
