+++ packages/espressocash_app/lib/features/activities/src/widgets/oskp_tile.dart
@@ -2,13 +2,13 @@ import 'package:auto_route/auto_route.dart';
 import 'package:dfunc/dfunc.dart';
 import 'package:flutter/material.dart';
 
-import '../../../../core/presentation/format_amount.dart';
 import '../../../../core/presentation/format_date.dart';
 import '../../../../gen/assets.gen.dart';
 import '../../../../l10n/device_locale.dart';
 import '../../../../l10n/l10n.dart';
 import '../../../../routes.gr.dart';
 import '../../../../ui/activity_tile.dart';
+import '../../../balances/presentation/format_amount.dart';
 import '../activity.dart';
 
 class OSKPTile extends StatelessWidget {
