+++ packages/espressocash_app/lib/features/activities/src/widgets/swap_tile.dart
@@ -2,7 +2,6 @@ import 'package:auto_route/auto_route.dart';
 import 'package:dfunc/dfunc.dart';
 import 'package:flutter/material.dart';
 
-import '../../../../core/presentation/format_amount.dart';
 import '../../../../core/presentation/format_date.dart';
 import '../../../../core/tokens/token.dart';
 import '../../../../l10n/device_locale.dart';
@@ -10,6 +9,7 @@ import '../../../../l10n/l10n.dart';
 import '../../../../routes.gr.dart';
 import '../../../../ui/activity_tile.dart';
 import '../../../../ui/token_icon.dart';
+import '../../../balances/presentation/format_amount.dart';
 import '../activity.dart';
 
 class SwapTile extends StatelessWidget {
