+++ packages/espressocash_app/lib/features/conversion_rates/conversion_rates.dart
@@ -1,0 +1,8 @@
+export 'src/bl/conversion_rates_bloc.dart'
+    show
+        ConversionRatesBloc,
+        ConversionRatesEvent,
+        ConversionRatesRequestException,
+        ConversionRatesState;
+export 'src/bl/repository.dart' show ConversionRatesRepository;
+export 'src/bl/token_price_cache.dart' show TokenPriceCacheRows;
