+++ packages/espressocash_app/lib/features/conversion_rates/module.dart
@@ -4,8 +4,8 @@ import 'package:nested/nested.dart';
 import 'package:provider/provider.dart';
 
 import '../../di.dart';
-import 'bl/conversion_rates_bloc.dart';
-import 'bl/repository.dart';
+import 'src/bl/conversion_rates_bloc.dart';
+import 'src/bl/repository.dart';
 
 class ConversionRatesModule extends SingleChildStatelessWidget {
   const ConversionRatesModule({Key? key, Widget? child})
