+++ packages/espressocash_app/lib/features/conversion_rates/src/bl/conversion_rates_client.dart
@@ -6,7 +6,7 @@ import 'package:freezed_annotation/freezed_annotation.dart';
 import 'package:injectable/injectable.dart';
 import 'package:retrofit/retrofit.dart';
 
-import '../../currency.dart';
+import '../../../../core/currency.dart';
 
 part 'conversion_rates_client.freezed.dart';
 part 'conversion_rates_client.g.dart';
