+++ packages/espressocash_app/lib/features/conversion_rates/src/bl/repository.dart
@@ -5,9 +5,9 @@ import 'package:flutter/material.dart';
 import 'package:injectable/injectable.dart';
 import 'package:rxdart/rxdart.dart';
 
-import '../../currency.dart';
-import '../../tokens/token.dart';
-import '../../tokens/token_list.dart';
+import '../../../../core/currency.dart';
+import '../../../../core/tokens/token.dart';
+import '../../../../core/tokens/token_list.dart';
 import 'conversion_rates_client.dart';
 import 'token_price_cache.dart';
