+++ packages/espressocash_app/lib/features/favorite_tokens/src/bl/bloc.dart
@@ -6,9 +6,9 @@ import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:freezed_annotation/freezed_annotation.dart';
 import 'package:injectable/injectable.dart';
 
-import '../../../../core/conversion_rates/bl/repository.dart';
 import '../../../../core/currency.dart';
 import '../../../../core/processing_state.dart';
+import '../../../conversion_rates/conversion_rates.dart';
 import 'repository.dart';
 
 part 'bloc.freezed.dart';
