+++ packages/espressocash_app/lib/features/favorite_tokens/widgets/favorite_tokens_list.dart
@@ -8,13 +8,13 @@ import '../../../../../core/tokens/token.dart';
 import '../../../../../routes.gr.dart';
 import '../../../../../ui/colors.dart';
 import '../../../../../ui/token_icon.dart';
-import '../../../core/conversion_rates/context_ext.dart';
-import '../../../core/presentation/extensions.dart';
 import '../../../core/user_preferences.dart';
 import '../../../di.dart';
 import '../../../l10n/device_locale.dart';
 import '../../../l10n/l10n.dart';
 import '../../../ui/theme.dart';
+import '../../balances/presentation/extensions.dart';
+import '../../conversion_rates/context_ext.dart';
 import '../src/bl/bloc.dart';
 import '../src/bl/repository.dart';
