+++ packages/espressocash_app/lib/features/fee/fee.dart
@@ -1,0 +1,1 @@
+export 'src/fee_label.dart';
