+++ packages/espressocash_app/lib/features/outgoing_direct_payments/src/widgets/odp_confirmation_screen.dart
@@ -4,7 +4,6 @@ import 'package:dfunc/dfunc.dart';
 import 'package:flutter/material.dart';
 import 'package:solana/solana.dart';
 
-import '../../../../core/fee_label.dart';
 import '../../../../core/tokens/token.dart';
 import '../../../../l10n/device_locale.dart';
 import '../../../../l10n/l10n.dart';
@@ -16,6 +15,7 @@ import '../../../../ui/button.dart';
 import '../../../../ui/dialogs.dart';
 import '../../../../ui/number_formatter.dart';
 import '../../../../ui/theme.dart';
+import '../../../fee/fee.dart';
 
 class ODPConfirmationScreen extends StatefulWidget {
   const ODPConfirmationScreen({
