+++ packages/espressocash_app/lib/features/outgoing_direct_payments/src/widgets/odp_details_screen.dart
@@ -2,7 +2,6 @@ import 'package:auto_route/auto_route.dart';
 import 'package:dfunc/dfunc.dart';
 import 'package:flutter/material.dart';
 
-import '../../../../core/presentation/format_amount.dart';
 import '../../../../core/presentation/format_date.dart';
 import '../../../../core/presentation/utils.dart';
 import '../../../../core/transactions/create_transaction_link.dart';
@@ -13,6 +12,7 @@ import '../../../../ui/timeline.dart';
 import '../../../../ui/transfer_status/transfer_error.dart';
 import '../../../../ui/transfer_status/transfer_progress.dart';
 import '../../../../ui/transfer_status/transfer_success.dart';
+import '../../../balances/presentation/format_amount.dart';
 import '../../models/outgoing_direct_payment.dart';
 import '../../widgets/extensions.dart';
 import '../bl/repository.dart';
