+++ packages/espressocash_app/lib/features/outgoing_split_key_payments/src/widgets/components/share_link.dart
@@ -2,13 +2,13 @@ import 'package:flutter/material.dart';
 import 'package:share/share.dart';
 
 import '../../../../../core/amount.dart';
-import '../../../../../core/presentation/format_amount.dart';
 import '../../../../../core/presentation/utils.dart';
 import '../../../../../l10n/l10n.dart';
 import '../../../../../ui/button.dart';
 import '../../../../../ui/colors.dart';
 import '../../../../../ui/share_message/share_message_bubble.dart';
 import '../../../../../ui/share_message/share_message_header.dart';
+import '../../../../balances/presentation/format_amount.dart';
 import '../../../models/outgoing_split_key_payment.dart';
 
 class ShareLinks extends StatelessWidget {
