+++ packages/espressocash_app/lib/features/payment_request/formatted_amount.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/widgets.dart';
 
-import '../../core/presentation/format_amount.dart';
 import '../../core/tokens/token_list.dart';
+import '../balances/presentation/format_amount.dart';
 import 'models/payment_request.dart';
 
 extension PaymentRequestFormatAmountExt on PaymentRequest {
