+++ packages/espressocash_app/lib/features/payment_request/src/widgets/link_details/components/share_link.dart
@@ -1,7 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:share/share.dart';
 
-import '../../../../../../core/presentation/format_amount.dart';
 import '../../../../../../core/presentation/utils.dart';
 import '../../../../../../core/tokens/token_list.dart';
 import '../../../../../../di.dart';
@@ -10,6 +9,7 @@ import '../../../../../../ui/button.dart';
 import '../../../../../../ui/colors.dart';
 import '../../../../../../ui/share_message/share_message_bubble.dart';
 import '../../../../../../ui/share_message/share_message_header.dart';
+import '../../../../../balances/presentation/format_amount.dart';
 import '../../../../models/payment_request.dart';
 
 class ShareLink extends StatelessWidget {
