+++ packages/espressocash_app/lib/features/swap/src/create_swap/create_swap_screen.dart
@@ -9,7 +9,6 @@ import '../../../../core/accounts/bl/account.dart';
 import '../../../../core/amount.dart';
 import '../../../../core/balances/bl/balances_bloc.dart';
 import '../../../../core/callback.dart';
-import '../../../../core/presentation/format_amount.dart';
 import '../../../../core/tokens/token.dart';
 import '../../../../di.dart';
 import '../../../../l10n/device_locale.dart';
@@ -19,6 +18,7 @@ import '../../../../ui/content_padding.dart';
 import '../../../../ui/dialogs.dart';
 import '../../../../ui/number_formatter.dart';
 import '../../../../ui/slider.dart';
+import '../../../balances/presentation/format_amount.dart';
 import '../../models/swap_seed.dart';
 import '../swap_operation.dart';
 import '../swap_route.dart';
