+++ packages/espressocash_app/lib/features/token_details/widgets/token_details_screen.dart
@@ -6,9 +6,6 @@ import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:provider/provider.dart';
 
 import '../../../core/amount.dart';
-import '../../../core/balances/presentation/watch_balance.dart';
-import '../../../core/presentation/extensions.dart';
-import '../../../core/presentation/format_amount.dart';
 import '../../../core/tokens/token.dart';
 import '../../../core/user_preferences.dart';
 import '../../../di.dart';
@@ -20,6 +17,9 @@ import '../../../ui/loader.dart';
 import '../../../ui/navigation_bar/navigation_bar.dart';
 import '../../../ui/theme.dart';
 import '../../../ui/token_icon.dart';
+import '../../balances/presentation/extensions.dart';
+import '../../balances/presentation/format_amount.dart';
+import '../../balances/presentation/watch_balance.dart';
 import '../../favorite_tokens/widgets/favorite_button.dart';
 import '../../ramp/widgets/ramp_buttons.dart';
 import '../../token_chart/module.dart';
