+++ packages/espressocash_app/test/core/conversion_rates/conversion_rates_provider_test.dart
@@ -1,9 +1,9 @@
 import 'package:decimal/decimal.dart';
-import 'package:espressocash_app/core/conversion_rates/bl/conversion_rates_client.dart';
-import 'package:espressocash_app/core/conversion_rates/bl/repository.dart';
-import 'package:espressocash_app/core/conversion_rates/bl/token_price_cache.dart';
 import 'package:espressocash_app/core/currency.dart';
 import 'package:espressocash_app/core/tokens/token.dart';
+import 'package:espressocash_app/features/conversion_rates/src/bl/conversion_rates_client.dart';
+import 'package:espressocash_app/features/conversion_rates/src/bl/repository.dart';
+import 'package:espressocash_app/features/conversion_rates/src/bl/token_price_cache.dart';
 import 'package:fast_immutable_collections/fast_immutable_collections.dart';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:mockito/annotations.dart';
