+++ packages/espressocash_app/test/core/presentation/format_amount_test.dart
@@ -1,8 +1,8 @@
 import 'dart:ui';
 
 import 'package:espressocash_app/core/amount.dart';
-import 'package:espressocash_app/core/presentation/format_amount.dart';
 import 'package:espressocash_app/core/tokens/token_list.dart';
+import 'package:espressocash_app/features/balances/presentation/format_amount.dart';
 import 'package:flutter_test/flutter_test.dart';
 
 void main() {
