+++ packages/espressocash_app/test/golden/golden_test.dart
@@ -4,12 +4,12 @@ import 'package:espressocash_app/core/accounts/bl/account.dart';
 import 'package:espressocash_app/core/accounts/bl/ec_wallet.dart';
 import 'package:espressocash_app/core/amount.dart';
 import 'package:espressocash_app/core/balances/bl/balances_bloc.dart';
-import 'package:espressocash_app/core/conversion_rates/bl/conversion_rates_bloc.dart';
-import 'package:espressocash_app/core/conversion_rates/bl/repository.dart';
 import 'package:espressocash_app/core/currency.dart';
 import 'package:espressocash_app/core/tokens/token.dart';
 import 'package:espressocash_app/core/user_preferences.dart';
 import 'package:espressocash_app/features/app_lock/src/bl/app_lock_bloc.dart';
+import 'package:espressocash_app/features/conversion_rates/src/bl/conversion_rates_bloc.dart';
+import 'package:espressocash_app/features/conversion_rates/src/bl/repository.dart';
 import 'package:espressocash_app/features/onboarding/src/widgets/no_email_and_password_screen.dart';
 import 'package:espressocash_app/features/sign_in/src/widgets/get_started_screen.dart';
 import 'package:flutter/material.dart';
